[Statepoint] Remove code related to inline operand bundles
This code becomes dead for valid IR after 8a0d6f2 and 1e29efc.  The reason for the test change is that the verifier reports the first verification error encountered, in some non-specified visit order.  By removing the verification code in gc.relocates for a statepoint with inline gc operands, I change the error the verifier reports.  And in one case, the checked for error is no longer possible with the bundle representation, so I simply delete the file.