AMDGPU: Don't sometimes allow instructions before lowered si_end_cf
Since b5683c02aa8c94ee0cbe009ebde0d6c712f43d4d, this would sometimes
not emit the or to exec at the beginning of the block, where it really
has to be. If there is an instruction that defines one of the source
operands, split the block and turn the si_end_cf into a terminator.

This avoids regressions when regalloc fast is switched to inserting
reloads at the beginning of the block, instead of spills at the end of
the block.

In a future change, this should always split the block.