[LoopDeletion] Remove over-eager SCEV verification.
20643804cbdcffa5280a4972de230b8ad8d2aa12 introduced SCEV verification to
deleteDeadLoop, but it appears this check is currently a bit over-eager
and some users of deleteDeadLoop appear to only patch up SE after
calling it (e.g. PR47753).

Remove the extra check for now. We can consider adding it back after we
tracked down the source of the inconsistency for PR47753.