Revert "[ARM][LowOverheadLoops] Adjust Start insertion."
This reverts commit 50ac3f1012833d0bd23f0212374cdccffe1b2652.

This commit contains some holes in its logic and has been causing
issues since it was commited. The idea sounds OK but some cases were not
handled correctly. Instead of trying to fix that up later it is probably
simpler to revert it and work to reimplement it in a more reliable way.