[DebugInfo] Follow up f72af0a1c4b6 with an API improvement
As mentioned post-commit in D85749, the 'substituteDebugValuesForInst'
method added in f72af0a1c4b6 would be better off with a limit on the
number of operands to substitute. This handles the common case of
"substitute the first operand between these two differing instructions",
or possibly up to N first operands.