[CostModel] remove cost-kind predicate for FP add/mul vector reduction costs
This was originally part of:
66dfcbf9fffe
but that was reverted because there was an underlying bug in
processing the vector type of these intrinsics. That was
fixed with:
a450d336fb94

This is similar in spirit to de74d15fd90e (memcpy) except that
here the underlying caller assumptions were created for vectorizer
use (throughput) rather than other passes.

That meant targets could have an enormous throughput cost with no
corresponding size, latency, or blended cost increase.

Paraphrasing from the previous commits:
This may not make sense for some callers, but at least now the
costs will be consistently wrong instead of mysteriously wrong.

Targets should provide better overrides if the current modeling
is not accurate.